Pull changes from route into master:
  Implement IPv4 as the first network layer protocol. The basics of the
  protocol are fully implemented:

    * IPv4 routing
    * IPv4 datagram forwarding
    * IPv4 fragmentation

  Additionally, various issue's with the networking core have been
  fixed (mostly memory leaks). PCap devices are now capable of a full
  shutdown, releasing all memory, to allow tools like valgrind to check
  for memory leaks.

From route:
9ee8e78 test: add a unit test for IP forwarding
eca7ad8 net: ip: implement IP forwarding
3b29145 net: dev: check weight before removing packet buffer
78348eb net: ip: implement IPv4 fragmentation
ba82d95 test: ip: add IP fragmentation unit test
734eb95 net: ip: implement IPv4 defragmentation
f497822 test: ip: update the IP unit test
2cb136a net: pcap: save incoming packets
b4c9543 net: ip: sent out protocol unreachable when required
5f8b3bb net: dev: poll all available devices
492c7fa ide: vscode: add the icmp unit test launch configuration
54f069e test: ip: add a unit test for the icmp protocol
582559b icmp: implement ICMP echo handling
2b6d241 net: dev: clear the RX flag when processing starts
6610aa4 test: ip: test the IP output routine
07bc1e5 net: ip: complete the IP output routine
1097faa ide: vscode: update file associations
28bdcac cmake: add relsize build type
c9984e9 ide: vscode: update the vscode IDE settings
8173086 estack: port: fix build issues for AVR target builds
859811e net: route: implement route deletion
aed31ef ide: vscode: update the workspace settings
37c540e net: routes: implement route lookups
3a49e82 net: dev: fix a typo in `netdev_print()'
3064a57 net: dev: let the netdev allocate the protocol handler
1336ca8 project: fix code style issue's
47593cd net: ip: add a setup for IP routing
dc4cd94 estack: ci: only run CI on the master branch

+++ source/addrutil.c
@@ -23,7 +23,7 @@ char *ethernet_mac_ntoa(uint8_t *addr, char *buf, size_t length)
 	assert(addr);
 	assert(buf);
 
-	if (length < 18)
+	if(length < 18)
 		return NULL;
 
 	snprintf(buf, length, "%02x:%02x:%02x:%02x:%02x:%02x",
@@ -53,7 +53,7 @@ uint32_t ipv4_atoi(char *addr)
 	uint32_t ip;
 
 	sscanf(addr, "%hhu.%hhu.%hhu.%hhu", &bytes[3], &bytes[2], &bytes[1], &bytes[0]);
-	ip = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[1] << 8) | bytes[0];
+	ip = ((uint32_t)bytes[3] << 24) | ((uint32_t)bytes[2] << 16) | ((uint32_t)bytes[1] << 8) | bytes[0];
 	return ip;
 }
